+++ apps/api/urls/rate_info_urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from ..views import rate_info_views as views
+
+app_name = 'Rate info endpoints'
+urlpatterns = [
+    path('read/', views.read_rate_info, name='read'),
+]